+++ podcasts/WatchManager.swift
@@ -204,6 +204,12 @@ class WatchManager: NSObject, WCSessionDelegate {
             if DateUtil.hasEnoughTimePassed(since: ServerSettings.lastRefreshEndTime(), time: 30.minutes) {
                 RefreshManager.shared.refreshPodcasts()
             }
+        } else if WatchConstants.Messages.LoginDetailsRequest.type == messageType {
+            // Watch is requesting login details but message was delivered without reply handler
+            // This can happen with WatchConnectivity timing issues
+            // so that updateWatchData() will push the latest login information to the watch.
+            FileLog.shared.addMessage("WatchManager: loginDetailsRequest received without reply handler, pushing data to watch")
+            updateWatchData()
         }
     }